[#11] feat: final

+++ src/main/java/com/nanum/board/boardservice/board/application/BoardServiceImpl.java
@@ -1,10 +1,7 @@
 package com.nanum.board.boardservice.board.application;
 
 import com.nanum.board.boardservice.board.domain.*;
-import com.nanum.board.boardservice.board.dto.BoardCategoryDto;
-import com.nanum.board.boardservice.board.dto.BoardDto;
-import com.nanum.board.boardservice.board.dto.BoardSearchCondition;
-import com.nanum.board.boardservice.board.dto.UserDto;
+import com.nanum.board.boardservice.board.dto.*;
 import com.nanum.board.boardservice.board.infrastructure.*;
 import com.nanum.board.boardservice.board.vo.*;
 import com.nanum.board.boardservice.client.UserServiceClient;
@@ -110,7 +107,7 @@ public class BoardServiceImpl implements BoardService {
         ModelMapper modelMapper = new ModelMapper();
         modelMapper.getConfiguration().setMatchingStrategy(MatchingStrategies.STRICT);
         BoardResponse boardResponse = modelMapper.map(board, BoardResponse.class);
-
+        boardResponse.setCategoryId(board.getBoardCategory().getId());
         if (id != -1L) {
             likeId = likeRepository.findByBoardIdAndUserId(postId, id);
             boardResponse.setProfileImgUrl(user.getResult().getProfileImgUrl());
@@ -183,7 +180,7 @@ public class BoardServiceImpl implements BoardService {
         Board board = boardRepository.findById(boardUpdateRequest.getBoardId()).get();
 
         List<Long> imgId = boardUpdateRequest.getImgId();
-
+        if(imgId !=null)
         imgId.forEach(boardImgRepository::deleteById);
 
         boardRepository.save(Board.builder()
@@ -198,7 +195,10 @@ public class BoardServiceImpl implements BoardService {
         BoardImage boardImages;
 
         S3UploadDto s3UploadDto;
-        if (multipartFiles != null && !multipartFiles.isEmpty()) {
+        if(multipartFiles==null){
+            return true;
+        }
+        if (!multipartFiles.isEmpty()) {
             for (MultipartFile multipartFile : multipartFiles) {
                 try {
                     s3UploadDto = s3UploaderService.uploadBoards(multipartFile, "myspharosbucket", "boardImg");
@@ -361,7 +361,7 @@ public class BoardServiceImpl implements BoardService {
                         .nestedCount(countNestReply)
                         .replyId(reply.getId())
                         .userId(reply.getUserId())
-                        .imgUrl(null)
+                        .imgUrl("/images/default.png")
                         .createAt(reply.getCreateAt())
                         .nickName(null)
                         .build();
@@ -385,6 +385,20 @@ public class BoardServiceImpl implements BoardService {
         return boardRepository.search(boardSearchCondition,pageable);
     }
 
+    @Override
+    public Page<BoardUserDto> findPostsByUser(Long userId, Pageable pageable) {
+      return  boardRepository.findAllByUserId(userId,pageable).map(board -> BoardUserDto.builder()
+                .id(board.getId())
+                .userId(board.getUserId())
+                .content(board.getContent())
+                .viewCount(board.getViewCount())
+                .title(board.getTitle())
+                .categoryId(board.getBoardCategory().getId())
+              .createAt(board.getCreateAt())
+                .build()
+        );
+    }
+
     //게시글 댓글들 조회
     @Override
     public List<ReplyResponse> retrieveReply(Long boardId) {
@@ -402,7 +416,7 @@ public class BoardServiceImpl implements BoardService {
                         .nestedCount(countNestReply)
                         .replyId(reply.getId())
                         .userId(reply.getUserId())
-                        .imgUrl(null)
+                        .imgUrl("/images/default.png")
                         .createAt(reply.getCreateAt())
                         .nickName(null)
                         .build());
@@ -475,9 +489,10 @@ public class BoardServiceImpl implements BoardService {
                         .content("삭제된 댓글입니다")
                         .nickName(null)
                         .userId(nestedReply.getUserId())
-                        .imgUrl(null)
+                        .imgUrl("/images/default.png")
                         .createAt(nestedReply.getCreateAt())
                         .replyId(nestedReply.getReply().getId())
+                        .id(nestedReply.getId())
                         .build());
             } else
                 replyResponses.add(ReplyResponse.builder()
@@ -487,6 +502,7 @@ public class BoardServiceImpl implements BoardService {
                         .imgUrl(users.getResult().getProfileImgUrl())
                         .createAt(nestedReply.getCreateAt())
                         .replyId(nestedReply.getReply().getId())
+                        .id(nestedReply.getId())
                         .build());
         });
         return replyResponses;
